Use rspec binstub if available.
Look for "bin/rspec" and use it, if present. This drastically improves RSpec’s start up time for projects using an application preloader like Spring (as long as the preloader generates a suitable binstub).

Note that the binstub is only used if there is a Gemfile present (otherwise it would not be possible to determine the RSpec version).