HTML output: Update filter while RSpec is running
Until now, if you changed the filter (eg. to hide  passed examples) while RSpec was still running, it only affected the examples that had be run so far. All examples that were completed after you changed the filter would be visible, regardless of the filter settings.

Todo: Get rid of the short flashing of examples before they get hidden (but this needs a bigger rewrite of the filter system, I’m afraid …)