Fix potential error when a spec fails without backtrace
The error could happen in projects using capybara-screenshot and a JavaScript driver like poltergeist: If a spec triggers an internal server error, it might have no backtrace associated, and then the formatter in capybara-screenshot would blow up because it doesn’t expect `#extra_failure_content` to return nil.

In RSpec’s HtmlFormatter (which is the “template” for our TextMateFormatter), `#extra_failure_content` always returns an array, even if there is no backtrace. So we should do the same.